+++ src/rer/solrpush/adapters/file.py
@@ -2,8 +2,6 @@
 from plone import api
 from rer.solrpush.interfaces.adapter import IExtractFileFromTika
 from zope.interface import implementer
-from ZODB.POSException import POSKeyError
-from ZODB.POSException import ReadConflictError
 
 try:
     from collective.limitfilesizepanel.interfaces import ILimitFileSizePanel
@@ -53,7 +51,9 @@ class FileExtractor(object):
             return None
         try:
             return file_obj.data
-        except (POSKeyError, ReadConflictError):
+        except Exception as e:
+            # generic because can be a PoskeyError raised also by RelStorage and we could not have it here.
             # this is raised if we are creating a new object, because there isn't the blob yet.
             # this is not a problem, because the file is in dictionary
+            logger.debug(e)
             return None